fix: carry structured output in result display item and harden review findings
Three fixes from code review:

1. Add structuredOutput field to DisplayItem so the SDK's structured_output
   (used for artifact materialization) survives the sidecar → frontend path.
   Previously outputText_result was always "Agent completed" and the actual
   payload was lost, causing steps 1/3 to error with "no structured output".

2. Change context_window from i32 to i64 in SidecarRunSummary to prevent
   silent serde deserialization failures for future large context windows.

3. Remove dead raw error pass-through from processErrorMessage (no frontend
   consumer after messages[] removal). Fix emitMockTurn to pass config
   context to MessageProcessor so mock streaming runs persist correctly.

+++ app/src-tauri/src/agents/events.rs
@@ -77,7 +77,7 @@ pub struct SidecarRunSummary {
     #[serde(rename = "modelUsageBreakdown")]
     pub model_usage_breakdown: Vec<SidecarModelUsageEntry>,
     #[serde(rename = "contextWindow")]
-    pub context_window: i32,
+    pub context_window: i64,
     #[serde(rename = "resultSubtype")]
     pub result_subtype: Option<String>,
     #[serde(rename = "resultErrors")]
